geAllUser for admin.. and roleBase routing.. and admin can't delete admin

+++ src/modules/user/user.service.ts
@@ -4,13 +4,10 @@ import type { UpdateUserDto } from "./dto/update-user-dto.js";
 export class UserServices{
     static async findAll() {
         try {
-            console.log('Executing query...'); 
             const result = await pool.query(`
                 SELECT id, name, email, phone, role 
                 FROM users
             `);
-            console.log('Query result:', result.rows.length); 
-            
             if(result.rows.length === 0){
                 throw new Error("No user found!");
             }
@@ -21,14 +18,20 @@ export class UserServices{
         }
     }
 
-      static async delete(userId: number) {
-            const existing = await pool.query("SELECT id FROM users WHERE id = $1", [userId]);
-            if (existing.rows.length === 0) {
-            throw new Error("User not found!");
-            }
-            await pool.query("DELETE FROM users WHERE id = $1", [userId]);
-            return true;
+    static async delete(userId: number) {
+        const existing = await pool.query("SELECT id, role FROM users WHERE id = $1", [userId]);
+      
+        if (existing.rows.length === 0) {
+          throw new Error("User not found!");
+        }
+        const userToDeleteRole = existing.rows[0].role;
+        if (userToDeleteRole === "admin") {
+          throw new Error("Admin Cannot delete Other admin!");
+        }
+        await pool.query("DELETE FROM users WHERE id = $1 AND role = $2", [userId, "customer"]);
+        return true;
       }
+      
 
 
       static async update(userId: number, dto: UpdateUserDto) {
